// src/app/model/tela-casaOracoes/validatorCasaOracoes.ts
export interface ValidatorCasaOracoes {
  igrId?: boolean;
  igrCod?: boolean;
  igrNome?: boolean;
  admNome?: boolean; // Incluindo admNome
  setorNome?: boolean;
  igrEstado?: boolean;
  igrCidade?: boolean;
  igrBairro?: boolean;
  igrCep?: boolean;
  igrEndereco?: boolean;
  igrComplemento?: boolean;
  admId?: boolean;
  setorId?: boolean;
}